Addressクラスで受け取れるように変更
