--- src/main/java/com/example/htmlpractice/Address.java
@@ -1,14 +1,0 @@
-package com.example.htmlpractice;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class Address {
-    String status;
-    String message;
-    List<Result> results;
-}
