+++ src/main/java/com/example/htmlpractice/AddressController.java
@@ -1,5 +1,6 @@
 package com.example.htmlpractice;
 
+import com.example.htmlpractice.model.Address;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
