--- src/main/java/com/example/htmlpractice/Result.java
@@ -1,18 +1,0 @@
-package com.example.htmlpractice;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class Result {
-    String zipcode;
-    String prefcode;
-    String address1;
-    String address2;
-    String address3;
-    String kana1;
-    String kana2;
-    String kana3;
-
-}
