+++ src/main/java/com/example/htmlpractice/model/Address.java
@@ -1,0 +1,17 @@
+package com.example.htmlpractice.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@Setter
+public class Address {
+    String message;
+    List<AddressData> results = new ArrayList<>();
+    int status;
+
+
+}
