+++ src/main/java/com/example/htmlpractice/model/AddressData.java
@@ -1,0 +1,13 @@
+package com.example.htmlpractice.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AddressData {
+    int zipcode;
+    String address1;
+    String address2;
+    String address3;
+}
